Move load_trades to bt_anlaysis
